requirements.txt
google-generativeai==0.8.3
httpx==0.27.2
gTTS==2.5.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
python-dotenv==1.0.1
praw==7.7.0
edge-tts>=7.0.0
instagrapi>=1.17.0
